fix for #266

+++ zipscript/include/race-file.h
@@ -1,13 +1,15 @@
 #ifndef _RACE_FILE_H_
 #define _RACE_FILE_H_
 
+#include <sys/stat.h>
+
 unsigned int	readsfv_file(struct LOCATIONS *, struct VARS *, int);
 void		delete_sfv_file(struct LOCATIONS *);
 void		readrace_file(struct LOCATIONS *, struct VARS *, struct USERINFO **, struct GROUPINFO **);
 void		maketempdir(struct LOCATIONS *);
 void		read_write_leader_file(struct LOCATIONS *, struct VARS *, struct USERINFO *);
 void		testfiles_file(struct LOCATIONS *, struct VARS *, int);
-int		copysfv_file(char *, char *, int);
+int		copysfv_file(char *, char *, off_t);
 void		create_indexfile_file(struct LOCATIONS *, struct VARS *, char *);
 short		clear_file_file(struct LOCATIONS *, char *);
 void		writerace_file(struct LOCATIONS *, struct VARS *, unsigned int, unsigned char);
